+++ backend/src/main/kotlin/pl/edu/agh/ki/covid19tablet/form/dto/FormDTO.kt
@@ -1,8 +1,8 @@
 package pl.edu.agh.ki.covid19tablet.form.dto
 
 import pl.edu.agh.ki.covid19tablet.form.FormId
-import pl.edu.agh.ki.covid19tablet.formState.dto.FormStateDTO
 import pl.edu.agh.ki.covid19tablet.schema.dto.SchemaDTO
+import pl.edu.agh.ki.covid19tablet.state.dto.FormStateDTO
 
 data class FormDTO(
     val id: FormId,
