--- backend/src/main/kotlin/pl/edu/agh/ki/covid19tablet/formStream/dto/update/ChoiceFieldStateUpdate.kt
@@ -1,8 +1,0 @@
-package pl.edu.agh.ki.covid19tablet.formStream.dto.update
-
-import pl.edu.agh.ki.covid19tablet.formState.fields.ChoiceFieldStateId
-
-data class ChoiceFieldStateUpdate(
-    val id: ChoiceFieldStateId,
-    val newValues: List<Boolean>
-)
