--- backend/src/main/kotlin/pl/edu/agh/ki/covid19tablet/formStream/dto/update/SignFieldStateUpdate.kt
@@ -1,8 +1,0 @@
-package pl.edu.agh.ki.covid19tablet.formStream.dto.update
-
-import pl.edu.agh.ki.covid19tablet.formState.fields.SignFieldStateId
-
-data class SignFieldStateUpdate(
-    val id: SignFieldStateId,
-    val newValue: String
-)
