--- backend/src/main/kotlin/pl/edu/agh/ki/covid19tablet/formStream/dto/update/SliderFieldStateUpdate.kt
@@ -1,8 +1,0 @@
-package pl.edu.agh.ki.covid19tablet.formStream.dto.update
-
-import pl.edu.agh.ki.covid19tablet.formState.fields.SliderFieldStateId
-
-data class SliderFieldStateUpdate(
-    val id: SliderFieldStateId,
-    val newValue: Double
-)
