+++ backend/src/main/kotlin/pl/edu/agh/ki/covid19tablet/state/fields/ChoiceFieldState.kt
@@ -1,7 +1,7 @@
-package pl.edu.agh.ki.covid19tablet.formState.fields
+package pl.edu.agh.ki.covid19tablet.state.fields
 
-import pl.edu.agh.ki.covid19tablet.formState.fields.dto.ChoiceFieldStateDTO
 import pl.edu.agh.ki.covid19tablet.schema.fields.ChoiceField
+import pl.edu.agh.ki.covid19tablet.state.fields.dto.ChoiceFieldStateDTO
 import javax.persistence.ElementCollection
 import javax.persistence.Entity
 import javax.persistence.GeneratedValue
