+++ backend/src/main/kotlin/pl/edu/agh/ki/covid19tablet/state/fields/SliderFieldState.kt
@@ -1,7 +1,7 @@
-package pl.edu.agh.ki.covid19tablet.formState.fields
+package pl.edu.agh.ki.covid19tablet.state.fields
 
-import pl.edu.agh.ki.covid19tablet.formState.fields.dto.SliderFieldStateDTO
 import pl.edu.agh.ki.covid19tablet.schema.fields.SliderField
+import pl.edu.agh.ki.covid19tablet.state.fields.dto.SliderFieldStateDTO
 import javax.persistence.Entity
 import javax.persistence.GeneratedValue
 import javax.persistence.Id
