+++ backend/src/main/kotlin/pl/edu/agh/ki/covid19tablet/state/fields/dto/ChoiceFieldStateDTO.kt
@@ -1,7 +1,7 @@
-package pl.edu.agh.ki.covid19tablet.formState.fields.dto
+package pl.edu.agh.ki.covid19tablet.state.fields.dto
 
-import pl.edu.agh.ki.covid19tablet.formState.fields.ChoiceFieldStateId
 import pl.edu.agh.ki.covid19tablet.schema.fields.ChoiceFieldId
+import pl.edu.agh.ki.covid19tablet.state.fields.ChoiceFieldStateId
 
 data class ChoiceFieldStateDTO(
     val id: ChoiceFieldStateId,
