+++ backend/src/main/kotlin/pl/edu/agh/ki/covid19tablet/state/fields/dto/SliderFieldStateDTO.kt
@@ -1,7 +1,7 @@
-package pl.edu.agh.ki.covid19tablet.formState.fields.dto
+package pl.edu.agh.ki.covid19tablet.state.fields.dto
 
-import pl.edu.agh.ki.covid19tablet.formState.fields.SliderFieldStateId
 import pl.edu.agh.ki.covid19tablet.schema.fields.SliderFieldId
+import pl.edu.agh.ki.covid19tablet.state.fields.SliderFieldStateId
 
 data class SliderFieldStateDTO(
     val id: SliderFieldStateId,
