+++ backend/src/main/kotlin/pl/edu/agh/ki/covid19tablet/state/fields/dto/TextFieldStateDTO.kt
@@ -1,7 +1,7 @@
-package pl.edu.agh.ki.covid19tablet.formState.fields.dto
+package pl.edu.agh.ki.covid19tablet.state.fields.dto
 
-import pl.edu.agh.ki.covid19tablet.formState.fields.TextFieldStateId
 import pl.edu.agh.ki.covid19tablet.schema.fields.TextFieldId
+import pl.edu.agh.ki.covid19tablet.state.fields.TextFieldStateId
 
 data class TextFieldStateDTO(
     val id: TextFieldStateId,
