+++ backend/src/main/kotlin/pl/edu/agh/ki/covid19tablet/stream/FormStreamController.kt
@@ -1,12 +1,12 @@
-package pl.edu.agh.ki.covid19tablet.formStream
+package pl.edu.agh.ki.covid19tablet.stream
 
 import org.springframework.messaging.handler.annotation.DestinationVariable
 import org.springframework.messaging.handler.annotation.MessageMapping
 import org.springframework.messaging.handler.annotation.SendTo
 import org.springframework.stereotype.Controller
 import pl.edu.agh.ki.covid19tablet.form.FormId
-import pl.edu.agh.ki.covid19tablet.formStream.dto.FormStateRequest
-import pl.edu.agh.ki.covid19tablet.formStream.dto.FormStateResponse
+import pl.edu.agh.ki.covid19tablet.stream.dto.FormStateRequest
+import pl.edu.agh.ki.covid19tablet.stream.dto.FormStateResponse
 
 @Controller
 class FormStreamController(
