+++ backend/src/main/kotlin/pl/edu/agh/ki/covid19tablet/stream/dto/update/ChoiceFieldStateUpdate.kt
@@ -1,0 +1,8 @@
+package pl.edu.agh.ki.covid19tablet.stream.dto.update
+
+import pl.edu.agh.ki.covid19tablet.state.fields.ChoiceFieldStateId
+
+data class ChoiceFieldStateUpdate(
+    val id: ChoiceFieldStateId,
+    val newValues: List<Boolean>
+)
