+++ backend/src/main/kotlin/pl/edu/agh/ki/covid19tablet/stream/dto/update/SliderFieldStateUpdate.kt
@@ -1,0 +1,8 @@
+package pl.edu.agh.ki.covid19tablet.stream.dto.update
+
+import pl.edu.agh.ki.covid19tablet.state.fields.SliderFieldStateId
+
+data class SliderFieldStateUpdate(
+    val id: SliderFieldStateId,
+    val newValue: Double
+)
